# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Visual Document Builder API
  version: 0.2.0
  description: |
    Canonical file — keep in sync with Nest routes.
    JWT issuance after OTP verify lands in P01-T02.
servers:
  - url: http://localhost:3001/api
    description: Local development
...
paths:
...
  /auth/otp/request:
    post:
      tags: [auth]
      summary: Request mobile OTP
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpBody'
      responses:
        '200':
          description: Challenge created (raw OTP never logged; optional devCode only in fake+development)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/RequestOtpResult'
        '400':
          $ref: '#/components/responses/ApiError'
        '429':
          $ref: '#/components/responses/ApiError'
        '503':
          $ref: '#/components/responses/ApiError'
  /auth/otp/verify:
    post:
      tags: [auth]
      summary: Verify OTP, upsert user, issue access JWT
      description: |
        Creates User on first success and returns a Bearer access token.
        No refresh token in MVP — client re-runs OTP when access expires.
        Logout blacklists token `jti` in Redis until expiry.
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpBody'
      responses:
        '200':
          description: User verified with access token
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/VerifyOtpResult'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
  /auth/login/options:
    post:
      tags: [auth]
      summary: Discover OTP / password / 2FA options for a mobile
      operationId: loginOptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpBody'
      responses:
        '200':
          description: Available login methods (no secrets)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/LoginOptions'
        '400':
          $ref: '#/components/responses/ApiError'
  /auth/password/login:
    post:
      tags: [auth]
      summary: Password login (or 2FA challenge when enabled)
      operationId: passwordLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginBody'
      responses:
        '200':
          description: Access token or requiresOtp challenge
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/VerifyOtpResult'
                      - $ref: '#/components/schemas/TwoFactorChallenge'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
  /auth/2fa/verify:
    post:
      tags: [auth]
      summary: Complete password + OTP two-factor login
      operationId: verifyTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorVerifyBody'
      responses:
        '200':
          description: Access token after 2FA
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/VerifyOtpResult'
        '401':
          $ref: '#/components/responses/ApiError'
  /auth/password:
    post:
      tags: [auth]
      summary: Set or change account password
      operationId: setPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordBody'
      responses:
        '200':
          description: Updated public user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/2fa:
    post:
      tags: [auth]
      summary: Enable or disable password+OTP two-factor
      operationId: setTwoFactor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorToggleBody'
      responses:
        '200':
          description: Updated public user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [auth]
      summary: Current user
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Public user profile
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [auth]
      summary: Update personal profile
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Updated public user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke current access token
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token jti blacklisted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [ok]
                    properties:
                      ok:
                        type: boolean
                        enum: [true]
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Phase 01-T02 — mobile OTP JWT access token
...
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: AUTH_REQUIRED
          description: |
            Identity codes include MOBILE_INVALID, OTP_COOLDOWN, OTP_RATE_LIMITED,
            OTP_INVALID, OTP_EXPIRED, REDIS_UNAVAILABLE, AUTH_REQUIRED, AUTH_INVALID
        message:
          type: string
          description: Fallback message; clients prefer i18n via code
    RequestOtpBody:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
          example: '09121234567'
    VerifyOtpBody:
      type: object
      required: [mobile, code]
      properties:
        mobile:
          type: string
          example: '09121234567'
        code:
          type: string
          pattern: '^\d{6}$'
          example: '123456'
    RequestOtpResult:
      type: object
      required: [mobile, expiresInSeconds, cooldownSeconds]
      properties:
        mobile:
          type: string
          example: '+989121234567'
        expiresInSeconds:
          type: integer
        cooldownSeconds:
          type: integer
        devCode:
          type: string
          description: Present only when SMS_PROVIDER=fake and NODE_ENV=development
    PublicUser:
      type: object
      required:
        [
          id,
          mobile,
          displayName,
          email,
          jobTitle,
          bio,
          trialConsumed,
          hasPassword,
          twoFactorEnabled,
          createdAt,
        ]
      properties:
        id:
          type: string
        mobile:
          type: string
        displayName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        jobTitle:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        trialConsumed:
          type: boolean
        hasPassword:
          type: boolean
          description: Whether a password is set (hash never exposed)
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
          maxLength: 120
        email:
          type: string
          nullable: true
          maxLength: 160
        jobTitle:
          type: string
          nullable: true
          maxLength: 120
        bio:
          type: string
          nullable: true
          maxLength: 1000
    LoginOptions:
      type: object
      required: [mobile, hasPassword, twoFactorEnabled, methods]
      properties:
        mobile:
          type: string
        hasPassword:
          type: boolean
        twoFactorEnabled:
          type: boolean
        methods:
          type: array
          items:
            type: string
            enum: [otp, password]
    PasswordLoginBody:
      type: object
      required: [mobile, password]
      properties:
        mobile:
          type: string
        password:
          type: string
          minLength: 8
    TwoFactorChallenge:
      type: object
      required: [requiresOtp, challengeToken, mobile, expiresInSeconds, cooldownSeconds]
      properties:
        requiresOtp:
          type: boolean
          enum: [true]
        challengeToken:
          type: string
        mobile:
          type: string
        expiresInSeconds:
          type: integer
        cooldownSeconds:
          type: integer
        devCode:
          type: string
    SetPasswordBody:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 8
        currentPassword:
          type: string
    TwoFactorToggleBody:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    TwoFactorVerifyBody:
      type: object
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
        code:
          type: string
          pattern: '^\d{4,8}$'
      type: object
      required: [id, name, role, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [OWNER, ADMIN, EDITOR, VIEWER]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VerifyOtpResult:
      type: object
      required: [user, isNewUser, accessToken, tokenType, expiresInSeconds]
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
        isNewUser:
          type: boolean
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresInSeconds:
          type: integer
...
  responses:
    ApiError:
      description: Error envelope
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/ApiError'
...